requirements.txt
python-telegram-bot==21.9
yt-dlp==2024.12.23
aiohttp==3.11.11